requirements.txt
aioschedule==0.5.2
discord.py==2.4.0
epicstore_api==0.1.7
Flask==3.0.3
psycopg2==2.9.9
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
Requests==2.32.3
